packages/api_python/requirements.txt
azure-functions>=1.17.0
azure-cosmos>=4.5.1
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.6.1
python-dotenv>=1.0.0
pypdf>=3.17.1
azure-core>=1.29.5
tiktoken>=0.5.2
tenacity>=8.2.3 